customer_service/requirements.txt
django==4.0.0
djangorestframework==3.13.0
djangorestframework-simplejwt==5.0.0
mysqlclient==2.1.0
pymysql==1.0.2
django-cors-headers==3.10.0
requests==2.26.0